+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * `getRequestsFromWarp10()`
   * `getAccessLogsHeatmapFromWarp10()`
   * `getAccessLogsDotmapFromWarp10()`
+* Add `withCache()` helper function
 
 ## 7.2.0 (2021-01-26)